backend/requirements.txt
# Backend Dependencies for Medical Chatbot

# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Data Validation
pydantic==2.5.0
pydantic-settings==2.1.0

# ChromaDB and Embeddings
chromadb==0.5.23
sentence-transformers==3.3.1
huggingface-hub==0.26.2

# HTTP Client
httpx>=0.27.0

# Environment Variables
python-dotenv==1.0.0

# Optional LLM Providers
openai==1.3.5  # Optional: for OpenAI API

# Utilities
python-json-logger==2.0.7
